Move packages to internal/ and use go:generate for bundling
