+++ internal/assets/assets.go
@@ -1,0 +1,6 @@
+package assets
+
+//go:generate fyne bundle -package assets -o bundled.go icon
+
+// AppIcon is the application icon for wormhole-gui
+var AppIcon = resourceIcon512Png
