+++ internal/assets/wormhole-gui.desktop
