+++ internal/bridge/bridge.go
