+++ internal/bridge/display.go
