+++ internal/bridge/receiver.go
