+++ internal/bridge/widgets/entry.go
