+++ internal/bridge/widgets/progress.go
